requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20
markdown==3.7
Pygments==2.19.0
pydantic==2.10.0
pydantic-settings==2.7.0
aiofiles==24.1.0
httpx==0.28.0
beautifulsoup4==4.12.0
python-slugify==8.0.0
pymupdf==1.25.0
python-docx==1.1.0